Include libXScrnSaver as a default dependency
Installing the RPM on a clean copy of the latest Fedora gives the
following error:

    /usr/share/atom/atom: error while loading shared libraries: libXss.so.1: cannot open shared object file: No such file or directory

I found a related issue that hit Atom itself[1]. The resolution was to
install libXScrnSaver. Rather than require all users to do this I
added it as a default dependency. This is the same solution Atom
implemented.

I did verify with libXScrnSaver now included as a requirement I can
now install the package and run it without error.

1. https://github.com/atom/atom/issues/13176